Demographic variables

Processes demographic variables.

clean_demographic(data)

Arguments

data
Data table - the Health Survey for England/Scotland Health Survey dataset.

Value

ethnicity_4cat: 4 level variable (see above).

ethnicity_2cat: 2 level variable (white/nonwhite).

sex: m/f

imd_quintile

Details

ETHNICITY In an attempt to harmonise different years of data to the receomended definitions, we have pooled the asian and other categories.

White (English, Irish, Scottish, Welsh, other European)

Mixed / multiple ethnic groups

Asian / Asian British (includes African-Indian, Indian, Pakistani, Bangladeshi), plus Other ethnic group (includes Chinese, Japanese, Philippino, Vietnamese, Arab)

Black / African / Caribbean / Black British (includes Caribbean, African)

Following inspection of the data, the white/non-white classification does look appropriate, especially given the likely limited sample sizes - so the 2 level variable has also been created.

For 2008-2013 of the Scottish Health Survey, we can create the same 4-category variable as for the HSE, howvere for 2014 onwards, the Scottish Health Survey 2018 only identifies 5 groups of ethnicity:

White (Scottish)

White (Other British)

White (Other)

Asian

Other minority ethnic

On the basis of this, only the 2 level variable (white/non-white) has been created for all years for Scotland.

Examples

if (FALSE) {

data_2001 <- read_2001()

data_2001 <- clean_demographic(data = data_2001)

}