hseclean: Health Survey Data Wrangling

Functions to read the data, rename, organise and process the variables required, impute missing data, and summarise data.

See also

Author

Maintainer: Duncan Gillespie duncan.gillespie@sheffield.ac.uk (ORCID)

Authors:

Laura Webster

Grace Leeming (ORCID)

Ryan KL Chen

Damon Morris (ORCID)

Colin Angus c.r.angus@sheffield.ac.uk

Alan Brennan a.brennan@sheffield.ac.uk