Read HSE 2012

Reads and does basic cleaning on the Health Survey for England 2012.

read_2012(
  root = c("X:/", "/Volumes/Shared/"),
  file = "ScHARR/PR_Consumption_TA/HSE/HSE 2012/UKDA-7480-tab/tab/hse2012ai.tab"
)

Arguments

root
Character - the root directory.

file
Character - the file path and name.

Value

Returns a data table. Note that:

Missing data ("NA", "", "-1", "-2", "-6", "-7", "-8", "-9", "-90", "-90.0", "N/A") is replaced with NA.

All variable names are converted to lower case.

The cluster and probabilistic sampling unit have the year appended to them.

Details

The HSE 2012 included a general population sample of adults and children, representative of the whole population at both national and regional level. For the sample, 9,024 addresses were randomly selected in 564 postcode sectors, issued over twelve months from January to December 2013. Where an address was found to have multiple dwelling units, one dwelling unit was selected at random and where there were multiple households at a dwelling unit, one household was selected at random.

A total of 8,291 adults aged 16 and over and 2,043 children aged 0-15 were interviewed. A household response rate of 64 population sample, 5,470 adults and 1,203 children had a nurse visit.

WEIGHTING

5.2 Individual weight

For analyses at the individual level, the weighting variable to use is (wt_int). These weights are generated separately for adults and children:

for adults (aged 16 or more), the interview weights are a combination of the householdweight and a component which adjusts the sample to reduce bias from individual non-response within households;

for children (aged 0 to 15), the weights are generated from the household weights and the child selection weights – the selection weights correct for only including a maximum of two children in a household. The combined household and child selection weight were adjusted to ensure that the weighted age/sex distribution matched that of all children in co-operating households.

For analysis of children aged 0-15 in both the Core and the Boost sample, taking into account child selection only and not adjusting for non-response, the (wt_child) variable can be used. For analysis of children aged 2-15 in the only Boost sample the (wt_childb) variable can

MISSING VALUES

-1 Not applicable: Used to signify that a particular variable did not apply to a given respondent usually because of internal routing. For example, men in women only questions.

-8 Don't know, Can't say.

-9 No answer/ Refused

Examples

if (FALSE) {

data_2012 <- read_2012("X:/", 
"ScHARR/PR_Consumption_TA/HSE/HSE 2012/UKDA-7480-tab/tab/hse2012ai.tab")

}