Read HSE 2017

Reads and does basic cleaning on the Health Survey for England 2017.

read_2017(
  root = c("X:/", "/Volumes/Shared/"),
  file =
    "HAR_PR/PR/Consumption_TA/HSE/Health Survey for England (HSE)/HSE 2017/UKDA-8488-tab/tab/hse17i_eul_v1.tab"
)

Arguments

root

Character - the root directory.

file

Character - the file path and name.

Value

Returns a data table. Note that:

Missing data ("NA", "", "-1", "-2", "-6", "-7", "-8", "-9", "-90", "-90.0", "N/A") is replaced with NA.

All variable names are converted to lower case.

The cluster and probabilistic sampling unit have the year appended to them.

Details

The HSE 2017 sample comprised of a core general population sample. There was no boost sample in 2017.

The sample comprised 9,612 addresses selected at random in 534 postcode sectors, issued over twelve months from January to December 2017. Field work was completed in March 2018.

Adults and children were interviewed at households identified at the selected addresses. Up to four children in each household were selected to take part at random; up to two aged 2 to 12 and up to two aged 13 to 15.

A total of 7,997 adults aged 16 and over and 1,985 children aged 0-15 were interviewed, including 5,196 adults and 1,195 children who had a nurse visit.#'

From 2015 HSE data contains the 2015 English index of multiple deprivation, divided into quintiles.

Examples

if (FALSE) {

data_2017 <- read_2017("X:/",
"ScHARR/PR_Consumption_TA/HSE/HSE 2017/UKDA-8334-tab/tab/hse2016_eul.tab")

}