Probabilistic mapping of Index of Multiple Deprivation to Townsend quintiles

A matrix that maps quintiles of the Index of Multiple Deprivation onto the Townsend Index of Deprviation.

imdq_to_townsend

Format

A data table

Source

See code in the data-raw folder of the hseclean package.

Details

To produce this We used area-level Office for National Statistics data to estimate the statistical association between the two metrics of deprivation, We used estimates of the Townsend Index from 2001 Census data at Ward level, and the Index of Multiple Deprivation 2015 (IMD 2015) at Lower-layer Super Output Area (LSOA) level. First, we mapped the 2001 definitions of Wards to the 2001 definitions of LSOAs. Second, we mapped the 2001 definitions of LSOAs to the 2011 definitions of LSOAs that are used by the IMD 2015.