Characteristics of smoking

Clean the variables that describe how much, what and to what level of addiction people smoke.

smk_amount(data)

Arguments

data
Data table - the Health Survey for England dataset.

Value

cigs_per_day - numeric (0+)

smoker_cat (non_smoker, 10_or_less, 11_to_20, 21_to_30, 31_or_more)

banded_consumption (non_smoker, light, moderate, heavy)

cig_type (non_smoker, hand rolled, machine rolled)

time_to_first_cig (non_smoker, less_than_5_minutes, five_to_thirty_minutes, thirty_minutes_but_less_than_1_hour, one_hour_or_more)

Details

The main variable is the average number of cigarettes smoked per day. For adults this is calculated from questions about how many cigarettes are smoked typically on a weekday vs. a weekend. For children, this is based on asking how many cigarettes were smoked in the last week. Missing values are imputed as the average amount smoked for an age, sex and IMD quintile subgroup.

We categorise cigarette preferences based on the answer to 'what is the main type of cigarette smoked'. In later years of the Health Survey for England, new questions are added that ask how many handrolled vs. machine rolled cigarettes are smoked on a weekday vs. a weekend. We currently don't use those questions because they were not asked in all years.

We also categorise the amount smoked, and use information on the time from waking until smoking the first cigarette. This latter variable has a high level of missingness. Together these categorical variables allow calculation of the heaviness of smoking index.

Examples

if (FALSE) {

data <- read_2001()
data <- clean_age(data)
data <- clean_demographic(data)
data <- smk_status(data)
data <- smk_amount(data)

}